Adding PuppetDB

+++ src/bash/agent.sh
@@ -5,6 +5,10 @@ export FACTER_repopath="$1"
 set -e
 set -x
 
-bash ${FACTER_repopath}/src/bash/repo.sh
+source ${FACTER_repopath}/src/bash/repo.sh
 
-apt-get install -y puppet
+if ! dpkg -l | grep -q 'ii  puppet '; then
+	apt-get install -y puppet
+fi
+
+source ${FACTER_repopath}/src/bash/fix-puppet.sh
